Make Wrapper validator executable and GitLab CI ready
Not all OS Projects are using GitHub and have the benefit of using
GitHub Actions. With this changes the Wrapper-Validation-Action
is executable on its own. It will use command line options for
configuration instead of GitHub Action Inputs.

Additionally a GitLab CI template was added and a Dockerfile for
building a ready to use Dockerimage.